Rename `isCroppingTop()`.
It's a getter for whether the "crop window" feature is enabled in settings; make the name communicate that more clearly.
Also rename the relevant `#define`s and variables and so on.

Along the way: tweak names from cropping "top" to general cropping (of any/all sides).  That's what I'm working on.