Remove PayPal button and homepage link resources.
It seems to be vestigial.  It looks like there's a space in the about dialog where it could have gone, but there's nothing in that space.  The link and button are referenced in a couple places in the code, but apparently only for event handling, without actually using/inserting them anywhere.
The homepage link is also 404 (in addition to not being visible anywhere I can find).