32-bit app icon
Like 5c9981005bbd0fa1ff18e073655dd07945ed2cc0 did for the tray icons, but for the primary app icon of the EXE itself.
I don't know the right way to do this (the "Condition" feature seems like it should be able to), so I've got to remember to manually change the icon, when making the 32 bit build.